+++ tools/macos-arp-helper/Sources/main.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+@main
+struct LanScannerMacOSHelper {
+    static func main() {
+        let args = Array(CommandLine.arguments.dropFirst())
+        CommandRouter.run(arguments: args)
+    }
+}